Refactor wgpu with stream
