+++ crates/cubecl-wgpu/src/compute/mod.rs
@@ -1,4 +1,7 @@
 pub(super) mod poll;
+pub(super) mod stream;
+pub(super) mod timestamps;
+
 mod server;
 mod storage;
